Question

+++ RealEstate.Api/AuthorizationService.cs
@@ -38,7 +38,7 @@ namespace RealEstate.Service.Classes
                     return;
                 }
                 var claims = (_principal.Claims).ToList();
-                Settings.Id  = Convert.ToInt64(claims.Where(c => c.Type == "Id").FirstOrDefault().Value);
+                Settings.Id  = Convert.ToInt32(claims.Where(c => c.Type == "Id").FirstOrDefault().Value);
                 Settings.Department = claims.Where(c => c.Type == "Department").FirstOrDefault().Value;
                 Settings.Name = claims.Where(c => c.Type == "Department").FirstOrDefault().Value;
                 context.HttpContext.User = _principal;
